update logging and move export into main and packages
