+++ Pipfile
@@ -13,3 +13,5 @@ aiohttp = ">2.3.0"
 idna-ssl = "==1.1.0"
 fire = "*"
 numpy = "*"
+tabulate = "*"
+structlog = "*"
