+++ packages/slack.py
@@ -1,8 +1,23 @@
 """Module to handle slack calling functions"""
-from time import sleep
-from requests import Session
+import asyncio
+import collections
+import os
+import time
 
-def upload_emoji(session:Session, emoji_name:str, filename:str):
+import aiohttp
+import requests
+
+from . import utils, log
+
+class SlackExportException(Exception):
+    """Class for slack export errors"""
+
+class NoEmojiException(Exception):
+    """Class for empty response errors"""
+
+Emoji = collections.namedtuple('Emoji', 'url name extension')
+
+def upload_emoji(session:requests.Session, emoji_name:str, filename:str):
     """uploads the emoji data to slack"""
     data = {
         'mode': 'data',
@@ -23,7 +38,7 @@ def upload_emoji(session:Session, emoji_name:str, filename:str):
             if resp.status_code == 429:
                 wait = int(resp.headers.get('retry-after', 1))
                 print(f"429 Too Many Requests!, sleeping for {wait} seconds")
-                sleep(wait)
+                time.sleep(wait)
                 continue
 
         resp.raise_for_status()
@@ -35,7 +50,7 @@ def upload_emoji(session:Session, emoji_name:str, filename:str):
 
         break
 
-def get_current_emoji_list(session:Session, all=False):
+def get_current_emoji_list(session:requests.Session, all_data=False):
     """List currently uploaded emoji to filter on"""
     page = 1
     result = []
@@ -50,7 +65,7 @@ def get_current_emoji_list(session:Session, all=False):
         resp.raise_for_status()
         response_json = resp.json()
 
-        if not all:
+        if not all_data:
             result.extend(map(lambda e: e["name"], response_json["emoji"]))
             if page >= response_json["paging"]["pages"]:
                 break
@@ -61,3 +76,104 @@ def get_current_emoji_list(session:Session, all=False):
 
         page = page + 1
     return result
+
+
+async def _determine_all_emoji_urls(session: aiohttp.ClientSession, base_url: str, token: str):
+    """pull all urls to download all emoji"""
+    logger = log.get_logger()
+    page = 1
+    total_pages = None
+
+    entries = list()
+
+    while total_pages is None or page <= total_pages:
+
+        data = {
+            'token': token,
+            'page': page,
+            'count': 100
+        }
+
+        response = await session.post(base_url + utils.EMOJI_API, data=data)
+
+        logger.info("loaded %s (page %d)", response.real_url, page)
+
+        if response.status != 200:
+            real_url = response.request_info.real_url
+            raise SlackExportException(
+                f"Failed to load emoji from {real_url} (status {response.status})")
+
+        json = await response.json()
+
+        for entry in json['emoji']:
+            url = str(entry['url'])
+            name = str(entry['name'])
+            extension = str(url.rsplit('.', maxsplit=1)[-1])
+
+            # slack uses 0/1 to represent false/true in the API
+            if entry['is_alias'] != 0:
+                logger.info(
+                    "Skipping emoji \"%s\", is alias of \"%s\"", name, entry['alias_for'])
+                continue
+
+            entries.append(Emoji(url, name, extension))
+
+        if total_pages is None:
+            total_pages = int(json['paging']['pages'])
+
+        page += 1
+
+    return entries
+
+
+
+def concurrent_http_get(max_concurrent: int, session: aiohttp.ClientSession):
+    """get emoji data"""
+    semaphore = asyncio.Semaphore(max_concurrent)
+
+    async def http_get(emoji: Emoji):
+        nonlocal semaphore
+        with await semaphore:
+            response = await session.get(emoji.url)
+            body = await response.content.read()
+            await response.wait_for_close()
+        return emoji, body
+
+    return http_get
+
+
+def save_to_file(response: bytes, emoji: Emoji, directory: str):
+    """save raw data to file"""
+    logger = log.get_logger()
+    logger.info("Downloaded %s from %s", emoji.name.ljust(20), emoji.url)
+    with open(os.path.join(directory, f"{emoji.name}.{emoji.extension}"), 'wb') as out:
+        out.write(response)
+
+
+async def export_loop():
+    """the async loop for exporting"""
+    logger = log.get_logger()
+    #TODO: update for args
+    args = None #args = _argparse()
+
+    if not os.path.exists(args.directory):
+        os.makedirs(args.directory)
+
+    base_url = utils.BASE_URL.format(team_name=args.team_name)
+
+    async with utils.async_session(args.cookie) as session:
+        #TODO: update for args
+        token = '' #await _fetch_api_token(session, base_url)
+
+        emojis = await _determine_all_emoji_urls(session, base_url, token)
+
+        if len(emojis) == 0:
+            raise NoEmojiException('Failed to find any custom emoji')
+
+        function_http_get = concurrent_http_get(args.concurrent_requests, session)
+
+        for future in asyncio.as_completed([function_http_get(emoji) for emoji in emojis]):
+            emoji, data = await future
+            save_to_file(data, emoji, args.directory)
+
+        logger.info("Exported %s custom emoji to directory '%s'", len(emojis), {args.directory})
